% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metis-tidy-package.R
\docType{package}
\encoding{UTF-8}
\name{metis.tidy}
\alias{metis.tidy}
\alias{metis.tidy-package}
\title{Access and Query Amazon Athena via the Tidyverse}
\description{
Methods are provided to use the 'metis' JDBC/DBI interface via
the Tidyverse (e.g. 'dbplyr'/'dplyr' idioms).
}
\references{
\href{https://s3.amazonaws.com/athena-downloads/drivers/JDBC/SimbaAthenaJDBC_2.0.6/docs/Simba+Athena+JDBC+Driver+Install+and+Configuration+Guide.pdf}{Simba Athena JDBC Driver with SQL Connector Installation and Configuration Guide}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
