% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_translate_env.R
\name{sql_translate_env.AthenaConnection}
\alias{sql_translate_env.AthenaConnection}
\title{Translate R tridyverse functional idioms to Athena}
\usage{
\method{sql_translate_env}{AthenaConnection}(con)
}
\arguments{
\item{con}{AthenaConnection}
}
\description{
Translate R tridyverse functional idioms to Athena
}
