% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_translate_env.R
\name{db_data_type.AthenaConnection}
\alias{db_data_type.AthenaConnection}
\alias{sql_translate_env.AthenaConnection}
\title{Convert R data type to Athena}
\usage{
db_data_type.AthenaConnection(con, fields, ...)

sql_translate_env.AthenaConnection(con)
}
\arguments{
\item{con}{Athena connection}

\item{fields}{fields to type reference}

\item{...}{ignored}

\item{con}{Athena connection}
}
\description{
Convert R data type to Athena

Translate R tridyverse functional idioms to Athena
}
